Plumb new Sunshine protocol extensions

+++ app/src/main/java/com/limelight/binding/input/driver/UsbDriverListener.java
@@ -1,7 +1,7 @@
 package com.limelight.binding.input.driver;
 
 public interface UsbDriverListener {
-    void reportControllerState(int controllerId, short buttonFlags,
+    void reportControllerState(int controllerId, int buttonFlags,
                                float leftStickX, float leftStickY,
                                float rightStickX, float rightStickY,
                                float leftTrigger, float rightTrigger);
